move to cmake, compilation done
