+++ .gitignore
@@ -1,0 +1,2 @@
+cmake-build-debug/*
+.idea/*
